изменение скриптов

+++ db/scripts/004_ddl_create_auto_post_table.sql
@@ -1,8 +1,8 @@
 create table auto_post (
-    id             serial primary key,
+    id             bigserial primary key,
     description    text not null,
     created        timestamp not null,
     sold           boolean not null,
     car_id         int references car(id),
-    user_id   int references auto_user(id) not null
+    user_id        int references auto_user(id) not null
 );
