+++ db/scripts/005_ddl_create_price_history_table.sql
@@ -1,7 +1,7 @@
 create table price_history (
-   id           serial primary key,
+   id           bigserial primary key,
    before       bigint not null,
    after        bigint not null,
    created      timestamp without time zone default now(),
-   post_id int references auto_post (id) not null
+   post_id      int references auto_post (id) not null
 );
